# Conflicts:
#	server/src/main/java/de/ingrid/igeserver/api/ConfigApiController.kt
#	server/src/main/java/de/ingrid/igeserver/services/ConnectionService.kt
#	server/src/main/java/de/ingrid/igeserver/services/SettingsService.kt

+++ server/src/main/java/de/ingrid/igeserver/exporter/model/SpatialModel.kt
@@ -1,6 +1,6 @@
 /**
  * ==================================================
- * Copyright (C) 2023-2024 wemove digital solutions GmbH
+ * Copyright (C) 2023-2025 wemove digital solutions GmbH
  * ==================================================
  * Licensed under the EUPL, Version 1.2 or – as soon they will be
  * approved by the European Commission - subsequent versions of the
@@ -28,7 +28,7 @@ data class SpatialModel(val type: String?, val title: String?, val value: Boundi
 
     val log = logger()
 
-    data class BoundingBoxModel(val lat1: Float, val lon1: Float, val lat2: Float, val lon2: Float)
+    data class BoundingBoxModel(val lat1: Double, val lon1: Double, val lat2: Double, val lon2: Double)
 
     val polygon: String?
         get() {
@@ -76,12 +76,11 @@ data class SpatialModel(val type: String?, val title: String?, val value: Boundi
         return null
     }
 
-    fun getWktCoordinatesISO(): String? =
-        if (this.wkt != null) {
-            convertWktToGml32(this.wkt)
-        } else {
-            null
-        }
+    fun getWktCoordinatesISO(): String? = if (this.wkt != null) {
+        convertWktToGml32(this.wkt)
+    } else {
+        null
+    }
 
     private fun getWktType(): String? {
         if (this.wkt != null) {
