+++ server/src/main/java/de/ingrid/igeserver/exports/iso/Decimal.kt
@@ -1,6 +1,6 @@
 /**
  * ==================================================
- * Copyright (C) 2023-2024 wemove digital solutions GmbH
+ * Copyright (C) 2023-2025 wemove digital solutions GmbH
  * ==================================================
  * Licensed under the EUPL, Version 1.2 or – as soon they will be
  * approved by the European Commission - subsequent versions of the
@@ -23,5 +23,5 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
 
 data class Decimal(
     @JacksonXmlProperty(localName = "Decimal", namespace = "http://www.isotc211.org/2005/gco")
-    var value: Float? = null,
+    var value: Double? = null,
 )
